Restructuring of files and folders
