+++ Application/Filter/IFilter.cs
@@ -1,0 +1,6 @@
+namespace Application.Filter;
+
+public interface IFilter<T>
+{
+    IQueryable<T> Filter(IQueryable<T> unfiltered, string? search);
+}
