--- Application/Interfaces/IMessage.cs
@@ -1,6 +1,0 @@
-namespace Application.Interfaces;
-
-public interface IMessage
-{
-
-}
