+++ Application/Messaging/IMessage.cs
@@ -1,0 +1,6 @@
+namespace Application.Messaging;
+
+public interface IMessage
+{
+
+}
