--- Application/Services/Stories/IStoriesService.cs
@@ -1,10 +1,0 @@
-using Application.Contracts;
-
-namespace Application.Services.Stories;
-
-public interface IStoriesService
-{
-    Task AddAsync(StoryDto storyDto);
-    Task<List<StoryDto>> GetAllAsync();
-    List<StoryDto> GetSortedStories(string? query);
-}
