--- Application/Services/Users/IUsersService.cs
@@ -1,15 +1,0 @@
-﻿using Application.Contracts;
-
-namespace Application.Services.Users
-{
-    public interface IUsersService
-    {
-        Task<List<UserDto>> GetAllAsync();
-        Task<UserDto> GetByEmailAsync(string email);
-        Task UpdateLastActiveAsync(string email);
-        Task AddInterestAsync(string email, InterestDto interest);
-        Task AddAsync(UserDto user);
-        Task DeleteInterestAsync(string email, int interestId);
-        Task<List<InterestDto>> GetInterestsAsync(string email);
-    }
-}
