+++ Application/Sort/ISorter.cs
@@ -1,0 +1,7 @@
+﻿namespace Application.Sort;
+
+public interface ISorter<T> where T : class
+{
+    List<T> Sort(IQueryable<T> unsorted, IEnumerable<SortingParameters> parameters);
+}
+
