+++ Application/Sort/SortField.cs
@@ -1,0 +1,6 @@
+namespace Application.Sort;
+
+public enum SortField
+{
+    None, Id, Score, Title
+}
