+++ Application/Sort/SortOrder.cs
@@ -1,0 +1,6 @@
+namespace Application.Sort;
+
+public enum SortOrder
+{
+    Asc, Desc
+}
