+++ Application/Sort/SortParameters.cs
@@ -1,0 +1,3 @@
+namespace Application.Sort;
+
+public record SortParameters(SortOrder Order, SortField FieldToSort);
