+++ Application/Stories/IStoriesRepository.cs
@@ -1,14 +1,14 @@
-using Application.Services;
+using Application.Sort;
 using Domain.Entities;
 
-namespace Application.Interfaces;
+namespace Application.Stories;
 
 public interface IStoriesRepository
 {
     Task<Story?> GetByIdAsync(int id);
     Task<List<Story>> GetByAuthorAsync(string author);
     Task<List<Story>> GetAllAsync();
-    List<Story> GetSortedStories(IEnumerable<SortingParameters> sortingParameters);
+    List<Story> GetAll(IEnumerable<SortingParameters> sortingParameters, string? search);
     Task AddAsync(Story story);
     Task UpdateAsync(int id, Story updatedStory);
     Task DeleteAsync(int id);
