+++ HackerNewsCommentsFeed/Controllers/StoriesController.cs
@@ -1,4 +1,4 @@
-using Application.Services.Stories;
+using Application.Stories;
 using Microsoft.AspNetCore.Mvc;
 using Shared.Utils;
 
@@ -9,11 +9,13 @@ public static class StoriesController
     public static IEndpointRouteBuilder MapItemsEndpoints(this IEndpointRouteBuilder app)
     {   
         app.MapGet("/api/stories", (
-            [FromQuery] string? orderBy, 
+            [FromQuery] string? orderBy,
+            [FromQuery] string? search,
             [FromServices] IStoriesService storiesService) =>
         {
-            var sortedStories = storiesService.GetSortedStories(orderBy);   
-            return Results.Ok(sortedStories);
+            var stories = storiesService.GetStories(orderBy, search);
+
+            return Results.Ok(stories);
             
         }).RequireAuthorization().WithTags(EndpointGroupTags.Stories);
