+++ Infrastructure/RabbitConnections/Subscriber/Subscriber.cs
@@ -1,6 +1,6 @@
 using System.Text;
 using System.Text.Json;
-using Application.Interfaces;
+using Application.Messaging;
 using Microsoft.Extensions.Logging;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
