+++ Shared/Utils/EndpointGroupTags.cs
@@ -1,11 +1,10 @@
-﻿namespace Shared.Utils
+﻿namespace Shared.Utils;
+
+public class EndpointGroupTags
 {
-    public class EndpointGroupTags
-    {
-        public const string Users = nameof(Users);
-        public const string Authentication = nameof(Authentication);
-        public const string Interests = nameof(Interests);
-        public const string Stories = nameof(Stories);
-        public const string UsersInterests = Users + " " + Interests;
-    }
+    public const string Users = nameof(Users);
+    public const string Authentication = nameof(Authentication);
+    public const string Interests = nameof(Interests);
+    public const string Stories = nameof(Stories);
+    public const string UsersInterests = Users + " " + Interests;
 }
