+++ Tests/Unit/SortingParametersParserTests.cs
@@ -1,4 +1,4 @@
-using Application.Services;
+using Application.Sort;
 using FluentAssertions;
 using Xunit;
