+++ Tests/Unit/UtilsTests.cs
@@ -1,7 +1,8 @@
 ﻿using System.Text.Json;
-using Application.Services;
+using Application.Sort;
 using Domain.Entities;
 using FluentAssertions;
+using Infrastructure.Db;
 using Shared.Utils;
 using Xunit;
